+++ changelog.md
@@ -17,7 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Bug` for any issue/error occurred
 - `Installed` for installing new dependencies
 - `Setting` for adding/changing the general settings
-          
+
+## [0.1.15] - 2022-01-15
+`Installed` webpack Merge 설정
+
+`Setting` webpack 설정(분리 => dev(devtool 변경),prod,common)
+
 ## [0.1.1] - 2022-01-09
 `Setting`
  - swagger api 문서 추가